feat: add scene selector and sidebar polish
