+++ src/components/Sidebar.css
@@ -59,6 +59,13 @@
   box-shadow: 0 28px 64px rgba(0,0,0,.45), inset 0 1px 0 rgba(255,255,255,.05);
   padding: 14px;
   box-sizing:border-box;
+  opacity: 0;
+  transform: translateX(-12px) scale(.96);
+  transition: transform var(--sb-duration) var(--sb-ease), opacity var(--sb-duration) var(--sb-ease);
+}
+.sb.open .sb-panel {
+  opacity: 1;
+  transform: none;
 }
 
 @media (max-width: 400px) {
@@ -173,3 +180,29 @@
   }
 }
 
+.world-select {
+  display: flex;
+  gap: 8px;
+  margin-top: 4px;
+}
+.world-option {
+  flex: 1;
+  padding: 6px 0;
+  border-radius: 12px;
+  border: 1px solid rgba(255,255,255,.16);
+  background: rgba(255,255,255,.08);
+  color: inherit;
+  cursor: pointer;
+  transition: background var(--sb-duration) var(--sb-ease), transform var(--sb-duration) var(--sb-ease);
+}
+.world-option:hover {
+  background: rgba(255,255,255,.16);
+}
+.world-option.on {
+  background: var(--accent);
+  color: #000;
+}
+.world-option:active {
+  transform: scale(.94);
+}
+.orb-slider { margin-top: 12px; }
