+++ src/hooks/useLocal.ts
@@ -1,0 +1,24 @@
+import { useEffect, useState } from "react";
+
+export function useLocal<T>(key: string, init: T) {
+  const [v, setV] = useState<T>(() => {
+    if (typeof window === "undefined") return init;
+    try {
+      const raw = window.localStorage.getItem(key);
+      return raw ? (JSON.parse(raw) as T) : init;
+    } catch {
+      return init;
+    }
+  });
+
+  useEffect(() => {
+    if (typeof window === "undefined") return;
+    try {
+      window.localStorage.setItem(key, JSON.stringify(v));
+    } catch {}
+  }, [key, v]);
+
+  return [v, setV] as const;
+}
+
+export default useLocal;
